refacrtor(ProjectRoutes): split struct and routes in two files
